Abilities begun
currently can add a light which is nice I suppose but what would be great would be a flare that you chuck and have on recharge that only last so long maybe. Also need a rope to spawn for scout really. Figure that out somehow, maybe a raycast or something